Adjust democracy proposals

+++ packages/api-derive/src/democracy/proposals.ts
@@ -56,7 +56,7 @@ export function proposals (instanceId: string, api: DeriveApi): () => Observable
             ? combineLatest([
               of(proposals),
               api.derive.democracy.preimages(
-                proposals.map(([, hash]) => hash)
+                proposals.map(([, hash]) => hash as Uint8Array)
               ),
               api.query.democracy.depositOf.multi(
                 proposals.map(([index]) => index)
